refactor: splitting types
