+++ ui/src/types/apiTypes.ts
@@ -1,0 +1,9 @@
+export interface APIResponse<T> {
+	success: boolean
+	message: string
+	data: T
+}
+export interface SourceTestResponse {
+	message: string
+	status: "FAILED" | "SUCCEEDED"
+}
