+++ ui/src/types/authTypes.ts
@@ -1,0 +1,4 @@
+export interface LoginArgs {
+	username: string
+	password: string
+}
