+++ ui/src/types/entityTypes.ts
@@ -1,0 +1,35 @@
+export interface Entity {
+	id: number
+	name: string
+	type: string
+	version: string
+	config: string
+	created_at: string
+	updated_at: string
+	created_by: string
+	updated_by: string
+	jobs: EntityJob[]
+}
+export interface EntityJob {
+	activate: boolean
+	destination_name?: string
+	source_name?: string
+	destination_type?: string
+	source_type?: string
+	id: number
+	job_name: string
+	last_run_state: string
+	last_runtime: string
+	name: string
+}
+export interface EntityBase {
+	name: string
+	type: string
+	version: string
+	config: string
+}
+export interface EntityTestRequest {
+	type: string
+	version: string
+	config: string
+}
